backend/requirements.txt
fastapi
uvicorn[standard]
python-multipart
pillow
opencv-python-headless
pytesseract
pdf2image
numpy
PyPDF2
pydantic
python-dotenv
google-cloud-documentai
google-auth
PyMuPDF
requests
difflib
